Refactor electron ipc listeners
