+++ electron/src/communication/ipc-events.ts
@@ -5,8 +5,8 @@ export default class IpcEvents {
         static readonly GET_IMG_NUM = 'get-images-number';
         static readonly GET_IMG_PAGE = 'get-images-page';
         static readonly MODIFY_EXIF = 'modify-exif';
-        static READ_SETTINGS = 'read-settings';
-        static SAVE_SETTINGS = 'save-settings';
+        static readonly READ_SETTINGS = 'read-settings';
+        static readonly SAVE_SETTINGS = 'save-settings';
     };
     static ToRendered = class {
         static readonly DIR_SELECTED = 'dir-selected';
@@ -14,6 +14,6 @@ export default class IpcEvents {
         static readonly IMG_NUM = 'images-number';
         static readonly IMG_PAGE_FOUND = 'images-page-found';
         static readonly MODIFY_EXIF_RESULT = 'modify-exif-result';
-        static SETTINGS_READY = 'settings-ready';
+        static readonly SETTINGS_READY = 'settings-ready';
     };
 }
