+++ electron/src/electron.dev.ts
@@ -1,6 +1,6 @@
 import { app, BrowserWindow, protocol } from 'electron';
 import ElectronConstants from './constants';
-import ipcCommunication from './ipc-communication';
+import { IpcCommunication } from './communication/ipc-communication';
 
 let window: BrowserWindow | null;
 
@@ -30,7 +30,8 @@ function createWindow(): void {
 }
 
 function loadIpcListeners(window: Electron.BrowserWindow) {
-    ipcCommunication.registerIpcHandlers(window);
+    const ipcCommunication = new IpcCommunication(window);
+    ipcCommunication.registerIpcHandlers();
 }
 
 app.whenReady().then(() => {
